third_party/blink/web_tests/platform/generic/fast/dom/StyleSheet/gc-parent-stylesheet-expected.txt

Test that CSS stylesheets don't lose custom properties on GC when only reachable through child rules.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.sheet is an instance of CSSStyleSheet
PASS style.sheet is null
PASS rule.parentStyleSheet.foo is "0"
PASS successfullyParsed is true

TEST COMPLETE
